fix: Internal error when saving debug logs with `--experimental=stateful-test-runner` or `--experimental=schema-analysis` enabled
Ref: #2353
