+++ src/schemathesis/service/serialization.py
@@ -3,14 +3,10 @@ from __future__ import annotations
 from dataclasses import asdict
 from typing import Any, Callable, Dict, Optional, TypeVar, cast
 
-from ..exceptions import format_exception
-from ..internal.result import Err, Ok
 from ..internal.transformation import merge_recursively
-from ..models import Response
 from ..runner import events
-from ..runner.serialization import SerializedCase, SerializedCheck
+from ..runner.serialization import _serialize_check
 from ..stateful import events as stateful_events
-from .models import AnalysisSuccess
 
 S = TypeVar("S", bound=events.ExecutionEvent)
 SerializeFunc = Callable[[S], Optional[Dict[str, Any]]]
@@ -38,17 +34,7 @@ def serialize_before_analysis(_: events.BeforeAnalysis) -> None:
 
 
 def serialize_after_analysis(event: events.AfterAnalysis) -> dict[str, Any] | None:
-    data = {}
-    analysis = event.analysis
-    if isinstance(analysis, Ok):
-        result = analysis.ok()
-        if isinstance(result, AnalysisSuccess):
-            data["analysis_id"] = result.id
-        else:
-            data["error"] = result.message
-    elif isinstance(analysis, Err):
-        data["error"] = format_exception(analysis.err())
-    return data
+    return event._serialize()
 
 
 def serialize_before_execution(event: events.BeforeExecution) -> dict[str, Any] | None:
@@ -56,50 +42,6 @@ def serialize_before_execution(event: events.BeforeExecution) -> dict[str, Any] 
         "correlation_id": event.correlation_id,
         "verbose_name": event.verbose_name,
         "data_generation_method": event.data_generation_method,
-    }
-
-
-def _serialize_case(case: SerializedCase) -> dict[str, Any]:
-    return {
-        "id": case.id,
-        "generation_time": case.generation_time,
-        "verbose_name": case.verbose_name,
-        "path_template": case.path_template,
-        "path_parameters": stringify_path_parameters(case.path_parameters),
-        "query": prepare_query(case.query),
-        "cookies": case.cookies,
-        "media_type": case.media_type,
-    }
-
-
-def _serialize_response(response: Response) -> dict[str, Any]:
-    return {
-        "status_code": response.status_code,
-        "headers": response.headers,
-        "body": response.body,
-        "encoding": response.encoding,
-        "elapsed": response.elapsed,
-    }
-
-
-def _serialize_check(check: SerializedCheck) -> dict[str, Any]:
-    return {
-        "name": check.name,
-        "value": check.value,
-        "request": {
-            "method": check.request.method,
-            "uri": check.request.uri,
-            "body": check.request.body,
-            "headers": check.request.headers,
-        },
-        "response": _serialize_response(check.response) if check.response is not None else None,
-        "example": _serialize_case(check.example),
-        "message": check.message,
-        "context": asdict(check.context) if check.context is not None else None,  # type: ignore
-        "history": [
-            {"case": _serialize_case(entry.case), "response": _serialize_response(entry.response)}
-            for entry in check.history
-        ],
     }
 
 
@@ -154,45 +96,7 @@ def serialize_stateful_event(event: events.StatefulEvent) -> dict[str, Any] | No
 
 
 def _serialize_stateful_event(event: stateful_events.StatefulEvent) -> dict[str, Any] | None:
-    data: dict[str, Any]
-    if isinstance(event, stateful_events.RunStarted):
-        data = {
-            "timestamp": event.timestamp,
-            "started_at": event.started_at,
-        }
-    elif isinstance(event, stateful_events.SuiteFinished):
-        data = {
-            "timestamp": event.timestamp,
-            "status": event.status,
-            "failures": [_serialize_check(SerializedCheck.from_check(failure)) for failure in event.failures],
-        }
-    elif isinstance(event, stateful_events.Errored):
-        data = {
-            "timestamp": event.timestamp,
-            "exception": format_exception(event.exception, True),
-        }
-    elif isinstance(event, stateful_events.StepFinished):
-        data = {
-            "timestamp": event.timestamp,
-            "status": event.status,
-            "transition_id": {
-                "name": event.transition_id.name,
-                "status_code": event.transition_id.status_code,
-                "source": event.transition_id.source,
-            }
-            if event.transition_id is not None
-            else None,
-            "target": event.target,
-            "response": {
-                "status_code": event.response.status_code,
-                "elapsed": event.response.elapsed.total_seconds(),
-            }
-            if event.response is not None
-            else None,
-        }
-    else:
-        data = asdict(event)
-    return {"data": {event.__class__.__name__: data}}
+    return {"data": {event.__class__.__name__: event.asdict()}}
 
 
 def serialize_after_stateful_execution(event: events.AfterStatefulExecution) -> dict[str, Any] | None:
@@ -264,28 +168,3 @@ def serialize_event(
             data = merge_recursively(data, extra)
     # Externally tagged structure
     return {event.__class__.__name__: data}
-
-
-def stringify_path_parameters(path_parameters: dict[str, Any] | None) -> dict[str, str]:
-    """Cast all path parameter values to strings.
-
-    Path parameter values may be of arbitrary type, but to display them properly they should be casted to strings.
-    """
-    return {key: str(value) for key, value in (path_parameters or {}).items()}
-
-
-def prepare_query(query: dict[str, Any] | None) -> dict[str, list[str]]:
-    """Convert all query values to list of strings.
-
-    Query parameters may be generated in different shapes, including integers, strings, list of strings, etc.
-    It can also be an object, if the schema contains an object, but `style` and `explode` combo is not applicable.
-    """
-
-    def to_list_of_strings(value: Any) -> list[str]:
-        if isinstance(value, list):
-            return list(map(str, value))
-        if isinstance(value, str):
-            return [value]
-        return [str(value)]
-
-    return {key: to_list_of_strings(value) for key, value in (query or {}).items()}
